feat: remve deprecated page, fix: lint errors
