+++ clients/desktop/src/vault/setup/onboarding/OnboardingController.tsx
@@ -1,10 +1,8 @@
-import { FEATURE_FLAGS } from '../../../constants'
 import { Match } from '../../../lib/ui/base/Match'
 import { useStepNavigation } from '../../../lib/ui/hooks/useStepNavigation'
 import { OnboardingGreeting } from './OnboardingGreeting'
 import { OnboardingSteps } from './OnboardingSteps'
 import { OnboardingSummary } from './OnboardingSummary'
-import { OnboardingSummaryOld } from './OnboardingSummaryOld'
 
 const steps = [
   'onboardingGreeting',
@@ -22,13 +20,7 @@ export const OnboardingController = () => {
         <OnboardingGreeting onCompleteGreeting={toNextStep} />
       )}
       onboardingSteps={() => <OnboardingSteps onCompleteSteps={toNextStep} />}
-      onboardingSummary={() =>
-        FEATURE_FLAGS.ENABLE_NEW_SUMMARY_PAGES ? (
-          <OnboardingSummary />
-        ) : (
-          <OnboardingSummaryOld />
-        )
-      }
+      onboardingSummary={() => <OnboardingSummary />}
     />
   )
 }
